requirements.txt
requests
googlesearch-python
pyfiglet
colorama

